Refactored some stuff, removed URL dependency because going to make custom network service
