+++ lib/frontend/components/event_page/event_header.dart
@@ -1,112 +1,73 @@
 import 'package:flutter/material.dart';
+import 'package:project_fomo/frontend/components/gradient_button.dart';
 import 'package:project_fomo/style.dart';
 
 class EventHeader extends StatelessWidget {
   final String imageURL;
-  final String artistName;
+  final String eventName;
 
-  EventHeader({@required this.imageURL, @required this.artistName});
+  EventHeader({@required this.imageURL, @required this.eventName});
 
   @override
   Widget build(BuildContext context) {
     return Stack(
-        alignment: Alignment.bottomLeft,
-        children: <Widget>[
-          Image.network(
-              imageURL,
+      alignment: Alignment.bottomLeft,
+      children: <Widget>[
+        Image.network(
+          imageURL,
           color: Color.fromRGBO(255, 255, 255, 0.5),
           colorBlendMode: BlendMode.modulate,
         ),
-          Padding(
-              padding: EdgeInsets.only(bottom: 10, left: 10),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
+        Padding(
+          padding: EdgeInsets.only(bottom: 10, left: 10),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: <Widget>[
+              Text(
+                eventName,
+                style: TextStyle(
+                    color: AppTextColor.title,
+                    fontWeight: AppFontWeight.bold,
+                    fontFamily: AppFontFamily.family,
+                    fontSize: AppFontSize.size28),
+              ),
+              Row(
                 children: <Widget>[
-                  Text(
-                      artistName,
+                  GradientButton(
+                    buttonText: Text(
+                      'Going',
                       style: TextStyle(
-                          color: AppTextColor.title,
-                          fontWeight: AppFontWeight.bold,
-                          fontFamily: AppFontFamily.family,
-                          fontSize: AppFontSize.size28
-                      )
+                        fontSize: AppFontSize.size18,
+                        fontFamily: AppFontFamily.family,
+                        color: AppTextColor.highEmphasis,
+                      ),
+                    ),
+                    buttonPressed: () {
+                      print('User wants to go');
+                    },
                   ),
-                  Row(
-                    children: <Widget>[
-                      GradientButton(
-                        buttonText: 'Going',
-                        buttonPressed: () {
-                          return null;
-                        },
+                  SizedBox(
+                    width: 15,
+                  ),
+                  GradientButton(
+                    buttonText: Text(
+                      'Interested',
+                      style: TextStyle(
+                        fontSize: AppFontSize.size18,
+                        fontFamily: AppFontFamily.family,
+                        color: AppTextColor.highEmphasis,
                       ),
-                      SizedBox(
-                        width: 15,
-                      ),
-                      GradientButton(
-                        buttonText: 'Interested',
-                        buttonPressed: () {
-                          return null;
-                        },
-                      )
-                    ],
-                  )
+                    ),
+                    buttonPressed: () {
+                      print('User is interested');
+                    },
+                  ),
                 ],
               )
-          )
-        ]
-    );
-
-
-  }
-}
-
-
-
-class GradientButton extends StatelessWidget {
-  final String buttonText;
-  final Function buttonPressed;
-
-  GradientButton({@required this.buttonText, @required this.buttonPressed});
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      margin: const EdgeInsets.only(top: 20.0),
-      child: Padding(
-        padding: const EdgeInsets.all(2.0),
-        child: Container(
-          child: ButtonTheme(
-              child: FlatButton(
-                color: Colors.transparent,
-                onPressed: buttonPressed,
-                shape: RoundedRectangleBorder(
-                  borderRadius: new BorderRadius.circular(50),
-                ),
-                child: Text(
-                  buttonText,
-                  textAlign: TextAlign.center,
-                  style: TextStyle(
-                    color: AppTextColor.highEmphasis,
-                    fontFamily: AppFontFamily.family,
-                    fontSize: AppFontSize.size22,
-                  ),
-                ),
-              ),
-              minWidth: 150.0,
-              height: MediaQuery.of(context).size.width * 0.12),
-          decoration: BoxDecoration(
-            color: AppColor.background,
-            border: Border.all(color: Colors.transparent),
-            borderRadius: BorderRadius.circular(50),
+            ],
           ),
         ),
-      ),
-      constraints: const BoxConstraints(minWidth: 88.0, minHeight: 36.0),
-      height: MediaQuery.of(context).size.width * 0.14,
-      decoration: BoxDecoration(
-        gradient: AppColor.gradient,
-        borderRadius: BorderRadius.circular(50),
-      ),
+      ],
     );
   }
 }
