+++ lib/frontend/components/event_page/info_line.dart
