+++ pubspec.yaml
@@ -8,7 +8,6 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  url_launcher: ^5.4.1
 
   cupertino_icons: ^0.1.2
   firebase_core: ^0.4.3+2
